pyproject.toml
[project]
name = "batch-review-mcp"
version = "0.1.0"
description = "Batch review tool for markdown files and code changes — MCP server + web UI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "fastmcp>=3.0.0",
    "gitpython>=3.1.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "websockets>=12.0",
]

[project.scripts]
batch-review = "main:cli_main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
# Include main.py at the root so the entry point script resolves correctly
sources = {"." = ""}
